run plugins onVisitNode during ast walking
